runs
it technically works, though curerntly the bet ive got is when you send a msg, it pulls all updates. would rather be live updates, but spent a lot of time trying to figure that out, and failed. i  think either adding another thread to check for updates, or use a gui. tell socket non-blocking? this was just suggested, mention it next class